requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-huggingface
sentence-transformers
faiss-cpu
python-dotenv
flask
pandas
streamlit

-e.
